Restructure into a proper Python package * normal python calling convention for main * specify poetry packages in pyproject.toml * nix flake uses poetry2nix to build the package * added .envrc for direnv * wrap rmapi in a more sane way with shutil.which
