+++ .envrc
@@ -1,0 +1,1 @@
+use flake
