+++ pyproject.toml
@@ -19,3 +19,6 @@ jinja2 = "^3.1.6"
 [build-system]
 requires = ["poetry-core"]
 build-backend = "poetry.core.masonry.api"
+
+[tool.poetry.scripts]
+zotero2remarkable-bridge = "zotero2remarkable_bridge.zotero2remarkable_bridge:main"
