+++ zotero2remarkable_bridge/sync_functions.py
@@ -8,7 +8,7 @@ from pprint import pprint
 
 from pyzotero.zotero import Zotero
 
-import rmapi_shim as rmapi
+import zotero2remarkable_bridge.rmapi_shim as rmapi
 import remarks
 from pathlib import Path
 from shutil import rmtree
